requirements.txt
matplotlib==3.10.6
numpy==2.3.3
OSMnx==2.0.6
pandas==2.3.2
python_igraph==0.11.9
